refactor: Path를 생성하는 로직을 Position에서 Path로 이동, 불필요한 getter 제거
